src/requirements.txt
SQLAlchemy
PyMySQL
python-dotenv
fastapi
uvicorn
openai
twilio
python-multipart
mercadopago